requirements.txt
# ─────────────────────────────────────────────
# Medical AI System – Python dependencies
# ─────────────────────────────────────────────

# API framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9   # required for FastAPI file uploads

# Deep learning
tensorflow==2.16.1        # includes Keras 3; use tensorflow-cpu if no GPU
# tensorflow-gpu==2.16.1  # uncomment for GPU-only builds

# Image processing
Pillow==10.3.0
opencv-python-headless==4.9.0.80   # headless = no GUI deps (Docker-friendly)
numpy==1.26.4

# ML utilities
scikit-learn==1.4.2       # metrics (classification_report, confusion_matrix)
matplotlib==3.8.4         # training curve plots

# Async & utilities
httpx==0.27.0             # async HTTP client (useful for testing)
python-dotenv==1.0.1      # env var management

# ─────────────────────────────────────────────
# Dev / testing (install with: pip install -r requirements-dev.txt)
# pytest==8.2.0
# pytest-asyncio==0.23.6
# httpx==0.27.0  (already above)
# ─────────────────────────────────────────────
tensorflow==2.16.1
numpy==1.26.4
protobuf==4.25.3
ml-dtypes==0.3.2
opencv-python==4.8.1.78
pillow
fastapi
uvicorn
python-multipart